+++ modules/nodes/04-k8s_nodes.tf
@@ -21,16 +21,6 @@ resource "aws_security_group_rule" "allow_all_egress_from_k8s_worker_nodes" {
   type              = "egress"
 }
 
-# FIXME: THIS NEED TO BE INJECTABLE BUT NOT HERE
-resource "aws_security_group_rule" "allow_all_from_us_workers" {
-  from_port         = 0
-  to_port           = 0
-  protocol          = "-1"
-  cidr_blocks       = [local.internal_network_cidr]
-  security_group_id = aws_security_group.k8s_workers_node_sg.id
-  type              = "ingress"
-}
-
 # Allow ALL connection from other nodes
 resource "aws_security_group_rule" "allow_all_from_self_workers" {
   from_port         = 0